+++ src/pages/Portal/index.js
@@ -13,9 +13,11 @@ import Widget from '../../molecules/Widget';
 
 import BoltArrowUpIcon from '../../atoms/BoltArrowUpIcon';
 import BoltArrowDownIcon from '../../atoms/BoltArrowDownIcon';
+import SunIcon from '../../atoms/SunIcon';
 
 import CurrentElectricityReceived from '../../molecules/CurrentElectricityReceived';
 import CurrentElectricityDelivered from '../../molecules/CurrentElectricityDelivered';
+import CurrentSolarPowerGenerating from '../../molecules/CurrentSolarPowerGenerating';
 
 export default function Portal() {
 	return (
@@ -30,6 +32,9 @@ export default function Portal() {
 				</Widget>
 				<Widget title="Delivering" name="current-electricity-delivered" icon={<BoltArrowUpIcon />}>
 					<CurrentElectricityDelivered />
+				</Widget>
+				<Widget title="Solar generating" name="current-solar-generating" icon={<SunIcon />}>
+					<CurrentSolarPowerGenerating />
 				</Widget>
 			</WidgetGrid>
 		</LoggedOut>