+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Loading page
+
+### Fixed
+- Portal logged out page.
+
 ## [1.21.1] - 2022-01-01
 ### Fixed
 - Wrong polling for all widgets, only polling for solar now.